Implement Category, Command, and CommandQueue
Category, Command, and CommandQueue are essential to the message passing
system in the scene graph. Scene nodes in the graph will have a virtual
getCategory() method, which by default passes the category stored on
construction.

Commands are pushed into the CommandQueue during update cycles. Commands
contain two properties, an Action and a Category. The action is a function
to be performed on a scene node that matches the category property.

In an upcoming commit, a portion of Overworld::update() involves popping
commands from its CommandQueue object until the queue is empty. The
commands are passed to the scene graph root's onCommand() method, and the
root will execute the command if it matches categories and then pass the
command to its children. The process continues recursively.