Add Player class
Player is much like the controller of MVC. It centralizes the input
handling for the GameState class. Player objects hold a reference to a
KeyBinding object and map the player actions to implementations that
execute the action. GameState will call the handleEvent() and
handleRealtimeInput() methods in its own handleEvent() method. It will
need to acquire a reference to the Overworld's command queue, which as of
this commit is not possible.

Player objects contain their own identifiers. The current implementation
passes the identifier to the action implementation, requesting that the
executing mob match identifiers. This implies that the identifier should
be interpreted as a player-controlled identifier, marking all the mobs
that the player sends actions to.