Override resolveCurrent() for Mob
Mobs resolve their collisions uniquely. If it's the mob's first step and
collides with something, it's position is reset to the original. If
however it is not the mob's first step and collides, then the mob
continues to move. This implies that if the mob successfully navigates its
first step, it has the privilege of completing its move even if another
mob attempts to access the same grid location.