Remove first step logic for Mob
For mobs to move from tile to tile, they originally had to pass the first
step test: If they could make their first step without colliding with
anything, then they were guaranteed to complete the tile move.

This however does not work if two mobs are approaching each other head-on.
In this case, both mobs successfully navigate their first steps, but the
collision would not resolve when they do come in contact. This results in
both mobs residing on the same tile -- undesired behavior!

Removing the first step logic will still mean only one of the mobs will
reset its position. This is determined arbitrarily though by its position
in the scene graph. A future commit may insure that the player mob is the
last added to the graph so that its moves are privileged above other mobs.