Add TMXParser class
Refactor Overworld code by extracting TMX parsing and placing into
specialized TMXParser class. The motivation for this is to adhere to the
Single Responsibility Principle by not muddying up the roles of the
classes. Originally, Overworld did the TMX file parsing and constructed
the vertices for TileMap to hold. Both these tasks fall outside the
intended purpose of the Overworld class, which is to manage the
lower-level components of an overworld, preferring delegation to internal
handling.

This commit has all parsing logic handled by TMXParser. TMXParser does not
construct vertices for TileMap, further separating responsibilities
appropriately. It instead returns the class-agnostic TileLayers type that
any class can handle as it sees fit. Overworld uses TMXParser to get a
TileLayers object, which it then passes to TileMap for it to construct map
vertices internally. This way TileMap can govern how it handles the tile
data passed to it.