Centralize layer logic in Overworld
Instead of Overworld containing one instance of TileMap that handles
several layers, Overworld contains an array of TileMap unique pointers
that represent the layers of the world. This keeps the implementation of
TileMap much simpler and centralizes the layer concept. Only Overworld has
any concept of a layer: SceneNode and TileMap are self-contained entities.

Overworld's scene graph is also organized into layers. Although there is
only one scene graph proper, the layer array points to specific nodes in
the graph that designate a layer. The layers are rendered by simply
iterating through the layer array.