Add LuaGameState class
Project has taken a new direction. For as much as can be done, gameplay
and higher level logic will be carried out in Lua scripts, while keeping
speed-critical code in C++ (especially the game loop). There are two main
advantages to this approach.

 - Less recompilation: When the time comes for level creation, it will be
   beneficial to tweak higher level details without having to recompile
   the code. If these details reside in Lua scripts, it can be accomodated
   easily.
 - Easier expression: The level of abstraction Lua provides will allow
   faster gameplay prototyping. C++ is of course very low-level where
   memory management and resource allocation are omnipresent concerns, but
   with Lua this is handled within the interpreter.

The Purple Pikachu project then should be regarded as an engine. It is fed
input in the form of Lua scripts and outputs to the screen. The game will
rest in the Lua scripts, powered by the C++ engine.

The most natural integration for Lua is within the State system. Lua uses
an environment state instance to interface with C++, and there can be
multiple Lua states for any C++ project. Since the TEMM::State classes
work independently of each other, a derived class LuaGameState can fit Lua
in nicely. LuaGameState wraps an instance of a Lua state and interfaces
with it through its static methods. Though they are static, they ought to
be interpreted as member methods. Object mapping is done through a static
registry of the LuaGameState class, since Lua cannot hold an instance to a
C++ object, only C functions.