Turn LuaGameState into template decorator class
LuaGameState (now LuaState) is now a template class that derives from the
State base class and wraps an arbitrary State derived/peer class (i.e,
decorator pattern). This allows LuaState to be plugged in anywhere a State
is expected, like StateStack, as well as extend the functionality of
existing State derived classes, which was the original purpose.

Traditional decorator patterns do not need templates. This is because the
decorator usually does not care about the type of derived class that it
decorates. In this situation, however, it is necessary that the decorator
LuaState know type information of the State class it wraps. This is
because LuaState provides additional functionality to specific State
derived classes, not just the general State base class.

Rather than create a new decorator class for each derived State class,
polymorphism is achieved through template specialization. The important
method is LuaState::exposeInterface(). It is specialized for any derived
State class that wants to expose functionality to Lua, and it is
automatically called in LuaState's constructor. All other methods are
defined as general methods for arbitrary wrapped States.*

The application at this commit initializes with a Lua wrapped GameState.
To wrap a new state in Lua, an additional boolean parameter is passed to
specify so. There is currently no functionality exposed to Lua.

* LuaState::exposeInterface() is defined generally as well. The program
  will default to this implementation if there is no specialized
  definition. It is a no-op.